Split out the tasks handlers and push adapters out
They now have their own module folder and moved the rest of the code
that didn't need to be in the toplevel of the project into the module
directory.
